Chatbot/RasaServer1/requirements.txt
flask
requests
rasa
pyyaml
flask_jwt_extended
requests
flask_sqlalchemy
pyotp